deduplicate constant definitions
Change the linux{major}_{minor} files that contain handwritten
bindings for constants that bindgen doesnt understand to be
"incremental" (e.g. linux4_14 only contains constants that are new to
4_14 and just reexports the 4_11 constants for old ones), to avoid
redefining an ever increasing set of constants with each version.

Also adjust the test that checks the api features and available ioctl
constants to only verify that the rust-defined set of features is a
subset of the kernel available features. For example running the test
on a 6.8 host kernel with the 5.7 feature means the kernel has more
features than the uffd crate supports, but this does not mean that the
crate definitions are _wrong_, just that they are incomplete.
