+++ userfaultfd-sys/Cargo.toml
@@ -19,4 +19,4 @@ cc = "1.0"
 [features]
 default = []
 linux4_14 = []
-linux5_7 = []
+linux5_7 = ["linux4_14"]
