+++ userfaultfd-sys/src/linux4_11.rs
@@ -14,6 +14,7 @@ pub const UFFD_API_FEATURES: u64 = UFFD_FEATURE_EVENT_FORK
 pub const UFFD_API_IOCTLS: u64 = 1 << _UFFDIO_REGISTER | 1 << _UFFDIO_UNREGISTER | 1 << _UFFDIO_API;
 pub const UFFD_API_RANGE_IOCTLS: u64 =
     1 << _UFFDIO_WAKE | 1 << _UFFDIO_COPY | 1 << _UFFDIO_ZEROPAGE;
+pub const UFFD_API_RANGE_IOCTLS_BASIC: u64 = 0;
 
 pub const UFFDIO_REGISTER_MODE_MISSING: u64 = 1 << 0;
 pub const UFFDIO_REGISTER_MODE_WP: u64 = 1 << 1;
@@ -36,7 +37,6 @@ mod const_tests {
 
     extern "C" {
         static _const_UFFD_API: u64;
-        static _const_UFFD_API_FEATURES: u64;
         static _const_UFFD_API_IOCTLS: u64;
         static _const_UFFD_API_RANGE_IOCTLS: u64;
         static _const_UFFDIO_REGISTER_MODE_MISSING: u64;
@@ -56,15 +56,7 @@ mod const_tests {
     fn consts_correct() {
         unsafe {
             assert_eq!(UFFD_API, _const_UFFD_API, "UFFD_API");
-            assert_eq!(
-                UFFD_API_FEATURES, _const_UFFD_API_FEATURES,
-                "UFFD_API_FEATURES"
-            );
             assert_eq!(UFFD_API_IOCTLS, _const_UFFD_API_IOCTLS, "UFFD_API_IOCTLS");
-            assert_eq!(
-                UFFD_API_RANGE_IOCTLS, _const_UFFD_API_RANGE_IOCTLS,
-                "UFFD_API_RANGE_IOCTLS"
-            );
             assert_eq!(
                 UFFDIO_REGISTER_MODE_MISSING, _const_UFFDIO_REGISTER_MODE_MISSING,
                 "UFFDIO_REGISTER_MODE_MISSING"
